+++ packages/webos/signLang/index.js
@@ -1,0 +1,17 @@
+/**
+ * Provide utility functions and higher order component for sign language.
+ *
+ * @module webos/signLang
+ * @exports SignLangDecorator
+ * @exports startSignLang
+ * @exports stopSignLang
+ */
+
+import {startSignLang, stopSignLang} from './signLang';
+import SignLangDecorator from './SignLangDecorator';
+
+export {
+	SignLangDecorator,
+	startSignLang,
+	stopSignLang
+};